Move source files into a directory
